refactoring

+++ src/athi_voxelgrid.h
@@ -11,13 +11,13 @@
 struct Node
 {
   Athi::Rect bounds;
-  std::vector<size_t> index;
+  std::vector<int> index;
   Node(const Athi::Rect &r);
-  void insert(size_t id);
-  void get(std::vector<std::vector<size_t>> &cont);
+  void insert(int id);
+  void get(std::vector<std::vector<int>> &cont);
   void draw() const;
   void color_objects();
-  bool contains(size_t id);
+  bool contains(int id);
 };
 
 class VoxelGrid
@@ -30,7 +30,7 @@ public:
   void init();
   void update();
   void draw() const;
-  void get(std::vector<std::vector<size_t>> &cont) const;
+  void get(std::vector<std::vector<int>> &cont) const;
 };
 
 extern VoxelGrid athi_voxelgrid;
@@ -38,7 +38,7 @@ extern VoxelGrid athi_voxelgrid;
 static void init_voxelgrid() { athi_voxelgrid.init(); }
 
 static void update_voxelgrid() { athi_voxelgrid.update(); }
-static void get_nodes_voxelgrid(std::vector<std::vector<size_t>> &cont)
+static void get_nodes_voxelgrid(std::vector<std::vector<int>> &cont)
 {
   athi_voxelgrid.get(cont);
 }
